Backend: a peer's chunk size cannot break the walk, and modify replaces
Both findings are on last round's code.

chunkedEnd added the peer's chunk size to the position of the chunk's first
byte in an int. parseChunkSize refuses only what overflows its own accumulator,
so "7fffffff" is a legal thing to receive, and after + chunk + 2 then wrapped
negative -- which read as "the chunk has not arrived yet" and put a negative
cursor back for the next pass. Measured with the fix removed: the walk indexes
the buffer at -2147483590 and the caller is handed a failure whose entire
message is that number. The review reading the same code predicted an endless
re-parse of the one size line, which is what it would be if indexOfCrLf did not
clamp its start; the code comment says both, because the arithmetic is what is
wrong either way. The Content-Length path had the same shape -- bodyStart +
declared could wrap and leave total negative, silently throwing away the
framing the loop had just worked out -- so both compute in a long now.

The kqueue modify added last round deleted only the filters the NEW event set
named, so changing READ to WRITE left EVFILT_READ armed: await() went on
reporting the descriptor readable for an interest the caller had replaced, and
since kevent reports one event per filter the same descriptor arrived twice in
one await. A modify replaces an interest -- EPOLL_CTL_MOD does it with its mask
and interestOps() does it on a Selector -- so both filters are dropped before
the requested ones are added.

+++ vm/backend/src/com/codename1/backend/Http.java
@@ -188,7 +188,12 @@ public final class Http {
         // length, or a chunk walk, or neither -- in which case end-of-stream is
         // the framing and this reads exactly as it used to.
         boolean framingKnown = false;
-        int total = -1;
+        // A LONG for the same reason chunkedEnd uses one: a peer may declare a
+        // Content-Length up to Integer.MAX_VALUE, and bodyStart + declared then
+        // wrapped negative -- which did not spin, but did leave total < 0, so the
+        // framing this loop had just worked out was silently ignored and the read
+        // went back to waiting for end of stream.
+        long total = -1;
         boolean chunked = false;
         int[] chunkCursor = new int[1];
         int frameStart = 0;
@@ -231,7 +236,7 @@ public final class Http {
                         chunkCursor[0] = bodyStart;
                     } else {
                         int declared = declaredLength(headNames, headValues);
-                        total = declared >= 0 ? bodyStart + declared : -1;
+                        total = declared >= 0 ? (long)bodyStart + (long)declared : -1;
                     }
                 }
             }
@@ -663,11 +668,25 @@ public final class Http {
                     scan = lineEnd + 2;
                 }
             }
-            int next = after + chunk + 2;   // the chunk's bytes and its own CRLF
+            // IN A WIDER TYPE, because the size is the PEER's number and it goes up
+            // to Integer.MAX_VALUE -- parseChunkSize refuses only what overflows
+            // its own accumulator. "7fffffff" made after + chunk + 2 wrap negative,
+            // which passed the bound below as though the chunk had already arrived
+            // and put a NEGATIVE cursor back for the next pass. Measured from
+            // there: the scan resumes at zero, because indexOfCrLf clamps its
+            // start, but `end` below begins at the unclamped cursor, so the walk
+            // indexes the buffer at -2147483590 and the caller is handed a failure
+            // whose whole message is that number. Without the clamp it is the
+            // endless re-parse of one size line instead. Neither is a reading of
+            // the peer's framing, and the arithmetic is what is wrong with both: a
+            // long cannot wrap here, and a chunk that big simply never fits, so
+            // this answers "not yet" until the read loop's response ceiling
+            // refuses it.
+            long next = (long)after + (long)chunk + 2L;   // the chunk and its CRLF
             if(next > size) {
                 return -1;
             }
-            pos = next;
+            pos = (int)next;
             cursor[0] = pos;
         }
     }
